adding prompt delete product
