+++ src/components/Navbar/navbar.module.css
@@ -65,6 +65,7 @@
     border-radius: 16px;
     height: 34px;
     padding: 0 16px;
+    cursor: pointer;
     border: none;
     color: var(--blanco);
 }
