+++ src/pages/Products/ProductView/productView.module.css
@@ -353,6 +353,60 @@ input[type=number]::-webkit-outer-spin-button {
 }
 
 
+.modal_delete_background{
+  position: fixed;
+  width: 100vw;
+  height: 100vh;
+  top: 0;
+  left: 0;
+  background-color: #000000b1;
+  display: flex;
+  align-items: center;
+  z-index: 1000;
+  justify-content: center;
+  animation: fade-in 0.4s ease-in;
+}
+
+.modal_delete_options{
+  display: flex;
+  flex-direction: column;
+  justify-content: space-around;
+  min-height: 200px;
+  background-color: var(--negro);
+  max-height: 350px;
+  width: 50%;
+  max-width: 450px;
+  border: 1px solid var(--blanco);
+  border-radius: 16px;
+  text-align: center;
+  min-width: 250px;
+  align-items: center;
+}
+.modal_delete_options> div{
+  display: flex;
+  justify-content: space-evenly;
+  width: 100%;
+}
+.modal_delete_options> div > button{
+  padding: 10px 5px;
+  border: none;
+  border-radius: 100px;
+  width: 90px;
+  font-size: 13px;
+  background-color: rgb(74, 78, 80);
+  color: white;
+  transition: background-color 0.2s;
+  cursor: pointer;
+}
+.modal_delete_options> div > button:hover{
+  background-color: var(--verde);
+
+}
+.modal_delete_options> div > button:last-child:hover{
+  background-color: rgb(240, 11, 11);
+
+}
+
 @media (min-width:468px){
   .product_info__title h1 {
     font-size: 24px;
